+++ mixins/sharedMethodsMixin.js
@@ -22,5 +22,9 @@ export default {
 		toggleRole(role) {
 			role.hover = !role.hover;
 		},
+		formatDate(dateString) {
+			const options = { year: 'numeric', month: 'long', day: 'numeric' };
+			return new Date(dateString).toLocaleDateString(undefined, options);
+		},
 	},
 };
