+++ pages/publications.vue
@@ -32,7 +32,7 @@
 						</v-tab>
 
 						<v-tab-item class="ma-5 mt-5 mt-md-0" :transition="false">
-							<Manuscripts :papers="papers.core"/>
+							<Manuscripts :papers="papers.core" />
 						</v-tab-item>
 						<v-tab-item class="ma-5 mt-5 mt-md-0" :transition="false">
 							<Manuscripts :papers="papers.collaboration" />
@@ -47,77 +47,7 @@
 				Posters
 			</v-tab>
 			<v-tab-item>
-				<v-card outlined class="pa-5" elevation="1">
-					<div v-if="selectedPoster">
-						<v-btn class="mb-4" @click="selectedPoster = null">
-							Back to posters
-						</v-btn>
-						<div class="selected-poster-details">
-							<embed
-								:src="getPosterPath(selectedPoster.poster)"
-								type="application/pdf"
-								width="100%"
-								height="600px"
-							/>
-							<h2>
-								<a :href="selectedPoster.link" target="_blank">{{
-									selectedPoster.title
-								}}</a>
-							</h2>
-							<p
-								v-if="
-									selectedPoster.authors && selectedPoster.authors.length > 0
-								"
-							>
-								<b>Authors:</b>
-								<span
-									v-for="(author, index) in selectedPoster.authors"
-									:key="index"
-								>
-									{{ author }}
-									<span v-if="index < selectedPoster.authors.length - 1"
-										>,
-									</span>
-								</span>
-							</p>
-							<p>
-								Published in {{ formatDate(selectedPoster.date) }}
-								{{ selectedPoster.publication_loc }}
-							</p>
-							<p>Abstract: {{ selectedPoster.abstract }}</p>
-							<p>Presented at {{ selectedPoster.presented_loc }}</p>
-						</div>
-					</div>
-					<div v-else>
-						<div class="license-text">
-							This is an open access work distributed under the terms of the
-							Creative Commons Attribution License, which permits unrestricted
-							use, distribution, and reproduction in any medium, provided the
-							original work is properly cited.
-						</div>
-						<div class="poster-grid">
-							<div
-								v-for="poster in posters"
-								:key="poster.title"
-								class="poster-preview"
-								@click="selectPoster(poster)"
-							>
-								<embed
-									:src="getPosterPath(poster.poster)"
-									type="application/pdf"
-									width="100%"
-									height="200px"
-								/>
-								<h3>
-									{{ poster.title }}
-								</h3>
-								<p class="publication-date">
-									{{ formatDate(poster.date) }}
-								</p>
-							</div>
-						</div>
-					</div>
-				</v-card>
+				<Posters :posters="posters"></Posters>
 			</v-tab-item>
 			<!-- <v-tab>
 				<v-icon left>mdi-school-outline</v-icon>
@@ -152,12 +82,14 @@
 <script>
 import posters from '@/static/posters/posters.json';
 import Manuscripts from '~/components/Cards/Manuscripts.vue';
+import Posters from '~/components/Cards/PosterList.vue';
 
 export default {
 	name: 'PublicationsPage',
 	components: {
-       Manuscripts
-    },
+		Manuscripts,
+		Posters,
+	},
 	data() {
 		return {
 			hostName: this.$config.OEB_LEGACY_ANGULAR_URI,
@@ -208,7 +140,6 @@ export default {
 	},
 	async mounted() {
 		await this.fetchAllPaperDetails();
-		this.posterDetails();
 	},
 	methods: {
 		async fetchPaperInfo(doi) {
@@ -248,48 +179,10 @@ export default {
 					}
 				}
 				// Sort papers by date after getting all details
-				this.papers[group].sort((a, b) => new Date(b.publicationDate) - new Date(a.publicationDate));
+				this.papers[group].sort(
+					(a, b) => new Date(b.publicationDate) - new Date(a.publicationDate)
+				);
 			}
-		},
-
-		getPosterPath(filename) {
-			return `${this.basePath}${filename}`;
-		},
-		selectPoster(poster) {
-			this.selectedPoster = poster;
-		},
-		posterDetails() {
-			this.posters.forEach((poster) => {
-				const filenameWithoutExt = poster.poster.split('.')[0];
-				const filePath = `@/static/posters/poster_list/${filenameWithoutExt}.json`;
-
-				import(filePath)
-					.then((module) => {
-						this.$set(poster, 'abstract', module.default.abstract);
-						this.$set(
-							poster,
-							'publication_loc',
-							module.default.publication_loc
-						);
-						this.$set(poster, 'link', module.default.link);
-						this.$set(poster, 'presented_loc', module.default.presented_loc);
-						this.$set(
-							poster,
-							'publicationDate',
-							module.default.publicationDate
-						);
-					})
-					.catch((error) => {
-						console.error(
-							`Error loading JSON for poster ${filenameWithoutExt}:`,
-							error
-						); // eslint-disable-line no-console
-					});
-			});
-		},
-		formatDate(dateString) {
-			const options = { year: 'numeric', month: 'long', day: 'numeric' };
-			return new Date(dateString).toLocaleDateString(undefined, options);
 		},
 	},
 };
@@ -309,64 +202,5 @@ export default {
 	width: 100%;
 	padding: 10px;
 	box-sizing: border-box;
-}
-
-.poster-grid {
-	display: flex;
-	flex-wrap: wrap;
-	gap: 30px;
-}
-
-.poster-preview {
-	width: calc(50% - 16px); /* Two posters per row */
-	cursor: pointer;
-	text-align: center;
-	border: 1px solid #ddd; /* Add border */
-	padding: 10px; /* Add padding */
-	background-color: #fff; /* Background color */
-	border-radius: 4px; /* Rounded corners */
-	display: flex;
-	flex-direction: column; /* Stack items vertically */
-	justify-content: space-between; /* Space out items to push date to bottom */
-}
-
-.poster-preview embed {
-	width: 100%;
-	height: 200px;
-}
-
-.poster-preview h3 {
-	margin: 10px 0;
-}
-
-.poster-preview .publication-date {
-	background-color: #f0f0f0; /* Background color for date */
-	padding: 10px;
-	border-radius: 4px;
-	width: 100%; /* Make it occupy full width */
-	text-align: center; /* Center the text */
-	margin-top: 10px;
-	width: 100%;
-	bottom: 0;
-	box-sizing: border-box;
-}
-
-.selected-poster-details {
-	text-align: left;
-}
-
-.selected-poster-details embed {
-	width: 100%;
-	height: 600px;
-	margin-bottom: 16px;
-}
-
-.license-text {
-	background-color: #f0f0f0; /* Adjust the background color as needed */
-	padding: 16px;
-	text-align: center;
-	margin-bottom: 20px;
-	border-radius: 4px;
-	font-weight: bold;
 }
 </style>
